eliminating warning

+++ imcmd_command_palette.cpp
@@ -682,7 +682,7 @@ void CommandPalette(const char* name, const char* hint)
     // Could be 0.5 on macOS Retina, 1 elsewhere
     float font_scale = ImGui::GetIO().FontGlobalScale;
 
-    if (gi.ExtraData.size() < item_count) {
+    if ((int)gi.ExtraData.size() < item_count) {
         gi.ExtraData.resize(item_count);
     }
